DedList
Password generator. Supply characters and password length, it will give every permutation with replacement. If a filename is supplied, it will output to that filename with extension ".txt" and will make more and more files (filename1.txt, filename2.txt...) if there are many passwords. If no filename it will print all passwords as it generates them so they can be piped into another program.  Number of character repititions can also be specified, but is default value=2.